+++ unscaled_to_scaled.cpp
@@ -9,7 +9,7 @@ void convertToUnscaled(const std::string& sampleName, const std::vector<float>& 
     std::cout<<"\n"; 
 }
 
-int main(uint8_t argc, char*argv[]) {
+int main(int argc, char*argv[]) {
     if(argc<2) error_handle(1); 
     std::string filePath = argv[1]; 
     BatchProcessor convertUnscaledtoScaled(filePath, convertToUnscaled);
